# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "This is an open API for `BAD` operations management. You can use the\
    \ api key `api_key` to test the authorization filters. <h2>Authentication</h2>\
    \ All API's must be authenticated. To authenticate, supply your API key as the\
    \ `Authorization` HTTP header with the `api_key` authorization scheme. <h2>Parameters</h2>\
    \ For API's that use POST, set the request's Content-Type header to `application/json`\
    \ and supply an appropriate JSON object in the request body. For API's that use\
    \ GET, specify each request parameter as a properly-escaped query parameter. <h2>Response</h2>\
    \ All responses have their Content-Type set to `application/json`, and contain\
    \ the response JSON object in the response body."
  version: "1.0.0"
  title: "BAD and Certified pick-ups OpenAPI"
host: "localhost:8081"
basePath: "/v2"
...
schemes:
- "http"
consumes:
- "application/json"
paths:
...
  /amp/create:
    post:
      tags:
      - "Amp"
      summary: "Create a new Amp"
      description: "Create a new Amp"
      operationId: "createAmp"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "amp object"
        required: true
        schema:
          $ref: "#/definitions/ampCreatePayload"
      - name: "initiator"
        in: "query"
        description: "Inintiator of the transaction"
        required: true
        type: "string"
      security:
      - api_key: []
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Message"
        "400":
          description: "Invalid amp supplied"
        "401":
          description: "Authentication information is missing or invalid"
        "404":
          description: "amp not found"
      x-swagger-router-controller: "Amp"
  /amp/update/{noAmp}:
    put:
      tags:
      - "Amp"
      summary: "Update an existing amp. Indicate values of attributes to update leave\
        \ the others empty"
      description: "Update basic details of an existing amp. A new token will be generated\
        \ if the details are modified."
      operationId: "updateAmp"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "noAmp"
        in: "path"
        description: "Amp number"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "updated amp object"
        required: true
        schema:
          $ref: "#/definitions/ampUpdatePayload"
      - name: "initiator"
        in: "query"
        description: "Inintiator of the transaction"
        required: true
        type: "string"
      security:
      - api_key: []
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Message"
        "400":
          description: "Invalid ID supplied"
        "401":
          description: "Authentication information is missing or invalid"
        "404":
          description: "amp not found"
      x-swagger-router-controller: "Amp"
  /amp/close/{noAmp}:
    put:
      tags:
      - "Amp"
      summary: "Close an Amp"
      description: "close an Amp"
      operationId: "closeAmp"
      produces:
      - "application/json"
      parameters:
      - name: "noAmp"
        in: "path"
        description: "Id of the Amp to close"
        required: true
        type: "string"
      - name: "initiator"
        in: "query"
        description: "Inintiator of the transaction"
        required: true
        type: "string"
      security:
      - api_key: []
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Message"
        "400":
          description: "Invalid ID supplied"
        "401":
          description: "Authentication information is missing or invalid"
        "404":
          description: "amp not found"
      x-swagger-router-controller: "Amp"
...
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
...
  ampCreatePayload:
    type: "object"
    required:
    - "declarantId"
    properties:
      noAmp:
        type: "string"
        example: "01E210621447"
      declarantId:
        type: "string"
        example: "PM0609"
      tractionExtNum:
        type: "string"
        example: "TRX15558556"
      version:
        type: "string"
        example: "1"
      noService:
        type: "string"
        example: "210621447"
      badId:
        type: "string"
        example: "01E210621447"
      statusBad:
        type: "string"
        example: "Forcé"
        enum:
        - "Non soumis"
        - "Soumis"
        - "Forcé"
        - "Rejeté"
        - "Transféré"
        - "Dissocié"
        - "Validé"
        - ""
      refEtatChargement:
        type: "string"
        example: "etat"
      transporterId:
        type: "string"
        example: "PM0582"
      tractorId:
        type: "string"
        example: "TRB537"
      unitId:
        type: "string"
        example: "YDGJB537"
      driver:
        type: "string"
        example: "driver"
      tag:
        type: "string"
        example: "tag"
      dateCreationAmpFret:
        type: "string"
        example: "01/02/2022"
    xml:
      name: "ampCreatePayload"
    description: "The inputs for creating a new amp"
  ampUpdatePayload:
    type: "object"
    required:
    - "ampStatus"
    - "declarantId"
    - "isTractorMandatory"
    - "noService"
    - "unitId"
    - "version"
    properties:
      noAmp:
        type: "string"
        example: "01E210621447"
      declarantId:
        type: "string"
        example: "PM0609"
      tractionExtNum:
        type: "string"
        example: "TRX15558556"
      version:
        type: "string"
        example: "1"
      noService:
        type: "string"
        example: "210621447"
      badId:
        type: "string"
        example: "01E210621447"
      statusBad:
        type: "string"
        example: "Forcé"
        enum:
        - "Non soumis"
        - "Soumis"
        - "Forcé"
        - "Rejeté"
        - "Transféré"
        - "Dissocié"
        - "Validé"
        - ""
      refEtatChargement:
        type: "string"
        example: "etat"
      transporterId:
        type: "string"
        example: "PM0582"
      tractorId:
        type: "string"
        example: "TRB537"
      isTractorMandatory:
        type: "boolean"
        example: true
        description: "Is the tractor mandatory, used only for the update of an amp"
      unitId:
        type: "string"
        example: "YDGJB537"
      driver:
        type: "string"
        example: "driver"
      tag:
        type: "string"
        example: "tag"
      dateCreationAmpFret:
        type: "string"
        example: "01/02/2022"
      ampStatus:
        type: "string"
        example: "Validée"
    xml:
      name: "ampUpdatePayload"
    description: "The inputs for updating updating an existing amp"
...
  Message:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
        minimum: 0
      message:
        type: "string"
      error:
        type: "boolean"
    example:
      code: 0
      error: true
      message: "Response Message"
...